+++ src/common/externalTranslationLoader.ts
@@ -1,7 +1,8 @@
 import {Injectable, Optional, Inject} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {isBlank, extend} from '@jscrpt/common';
 import {HTTP_REQUEST_BASE_URL} from '@anglr/common';
+import {isBlank} from '@jscrpt/common';
+import {extend} from '@jscrpt/common/extend';
 import {TranslateLoader} from '@ngx-translate/core';
 import {Observable, Observer, forkJoin} from 'rxjs';
 
@@ -10,7 +11,7 @@ import {ExternalTranslationLoaderOptions} from './externalTranslationLoaderOptio
 /**
  * External translation loader, that can be configured with multiple resources
  */
-@Injectable()   
+@Injectable()
 export class ExternalTranslationLoader implements TranslateLoader
 {
     //######################### private fields #########################
@@ -27,17 +28,17 @@ export class ExternalTranslationLoader implements TranslateLoader
     {
         if(isBlank(_baseUrl))
         {
-            this._baseUrl = "";
+            this._baseUrl = '';
         }
 
         if(!_options || !(_options instanceof ExternalTranslationLoaderOptions))
         {
-            this._options = new ExternalTranslationLoaderOptions("translations", ["lang"], ".json");
+            this._options = new ExternalTranslationLoaderOptions('translations', ['lang'], '.json');
         }
     }
 
     //######################### public methods #########################
-    
+
     /**
      * Gets translations for language
      * @param lang - Id of language
@@ -46,11 +47,11 @@ export class ExternalTranslationLoader implements TranslateLoader
     {
         return Observable.create((observer: Observer<any>) =>
         {
-            var translationsResources: Promise<any>[] = [];
-            
+            const translationsResources: Promise<any>[] = [];
+
             this._options.resources.forEach(itm =>
             {
-                let url = `${this._baseUrl}${this._options.resourcePrefix}/${lang}/${itm}${this._options.resourceSufix}`;
+                const url = `${this._baseUrl}${this._options.resourcePrefix}/${lang}/${itm}${this._options.resourceSufix}`;
 
                 if(this._cachedResults[url])
                 {
@@ -68,17 +69,17 @@ export class ExternalTranslationLoader implements TranslateLoader
                     translationsResources.push(this._cachedResults[url]);
                 }
             });
-            
+
             forkJoin(translationsResources)
                 .subscribe(success =>
                            {
-                               var translations = {};
-                               
-                               for(var index in success)
+                               const translations = {};
+
+                               for(const index in success)
                                {
                                    extend(translations, success[index]);
                                }
-                               
+
                                observer.next(translations);
                                observer.complete();
                            },